Remove unused invite code logic
