+++ LocalPackages/DataBrokerProtection/Tests/DataBrokerProtectionTests/Mocks.swift
@@ -428,96 +428,6 @@ final class CaptchaServiceMock: CaptchaServiceProtocol {
     }
 }
 
-final class MockRedeemUseCase: DataBrokerProtectionRedeemUseCase {
-    var shouldSendNilAuthHeader = false
-
-    func getAuthHeader() -> String? {
-        if shouldSendNilAuthHeader {
-            return nil
-        }
-        return "auth header"
-    }
-
-    func shouldAskForInviteCode() -> Bool {
-        false
-    }
-
-    func redeem(inviteCode: String) async throws {
-
-    }
-
-    func reset() {
-        shouldSendNilAuthHeader = false
-    }
-}
-
-final class MockAuthenticationService: DataBrokerProtectionAuthenticationService {
-
-    var wasRedeemCalled = false
-    var shouldThrow = false
-
-    func redeem(inviteCode: String) async throws -> String {
-        wasRedeemCalled = true
-        if shouldThrow {
-            throw AuthenticationError.issueRedeemingInviteCode(error: "mock")
-        }
-
-        return "accessToken"
-    }
-
-    func reset() {
-        wasRedeemCalled = false
-        shouldThrow = false
-    }
-}
-
-final class MockAuthenticationRepository: AuthenticationRepository {
-    var shouldSendNilInviteCode = false
-    var shouldSendNilAccessToken = false
-    var wasInviteCodeSaveCalled = false
-    var wasAccessTokenSaveCalled = false
-    var shouldSendNilWaitlistTimeStamp = false
-
-    func getInviteCode() -> String? {
-        if shouldSendNilInviteCode {
-            return nil
-        }
-
-        return "inviteCode"
-    }
-
-    func getAccessToken() -> String? {
-        if shouldSendNilAccessToken {
-            return nil
-        }
-
-        return "accessToken"
-    }
-
-    func save(inviteCode: String) {
-        wasInviteCodeSaveCalled = true
-    }
-
-    func save(accessToken: String) {
-        wasAccessTokenSaveCalled = true
-    }
-
-    func getWaitlistTimestamp() -> Int? {
-        if shouldSendNilWaitlistTimeStamp {
-            return nil
-        }
-        return 123
-    }
-
-    func reset() {
-        shouldSendNilInviteCode = false
-        shouldSendNilAccessToken = false
-        wasInviteCodeSaveCalled = false
-        wasAccessTokenSaveCalled = false
-        shouldSendNilWaitlistTimeStamp = false
-    }
-}
-
 final class BrokerUpdaterRepositoryMock: BrokerUpdaterRepository {
     var wasSaveLatestAppVersionCheckCalled = false
     var lastCheckedVersion: String?
